requirements.txt
pandas
numpy
scikit-learn
xgboost
plotly
streamlit
PyYAML
